requirements.txt
azure-storage-blob==12.20.0
black==24.10.0
dash-bootstrap-components==1.6.0
dash-extensions==1.0.18
dash-leaflet==1.0.15
dash-mantine-components==0.12.1
flake8==7.1.1
geopandas==1.0.1
gunicorn==22.0.0
importlib-resources==6.4.0
isort==5.13.2
pip-chill==1.0.3
pre-commit==4.0.1
psycopg2-binary==2.9.10
pyarrow==16.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
tomli==2.0.1
jenkspy
